requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.5
aiohttp-jinja2==1.6
aioopenssl==0.6.0
aiosasl==0.5.0
aiosignal==1.3.1
aioxmpp==0.13.3
arrow==1.3.0
attrs==24.2.0
babel==2.16.0
cffi==1.17.1
cryptography==43.0.3
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
Jinja2==3.0.3
jinja2-time==0.2.0
lxml==4.9.4
MarkupSafe==3.0.2
multidict==5.2.0
propcache==0.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2022.1
singletonify==0.2.4
six==1.16.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
spade==3.3.3
timeago==1.0.16
types-python-dateutil==2.9.0.20241003
tzlocal==5.2
yarl==1.18.0

PyYAML~=6.0.2